requirements.txt
asgiref==3.7.2
Django==5.0.2
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-mptt==0.16.0
pillow==10.2.0
sqlparse==0.4.4
tzdata==2024.1
